requirements.txt
pyyaml
pytest
selenium
pytesseract
requests
google_auth_oauthlib
google-api-python-client
fake_useragent
opencv-python